requirements.txt
amqp
asgiref
billiard
black
celery
certifi
channels
charset-normalizer
click
click-didyoumean
click-plugins
click-repl
Django
django-allauth
django-axes
django-channels
django-debug-toolbar
django-filter
django-guardian
django-parler
django-recaptcha
django-rest-knox
django-taggit
djangochannelsrestframework
djangorestframework
drf-yasg
factory_boy
Faker
gunicorn
idna
inflection
kombu
mypy-extensions
oauthlib
packaging
pathspec
pillow
platformdirs
prompt_toolkit
psycopg2-binary
python-dateutil
python-decouple
pytz
PyYAML
redis
requests
requests-oauthlib
six
sqlparse
typing_extensions
tzdata
uritemplate
vine
urllib3
wcwidth
daphne
channels-redis
pytest-mock
pytest-drf
pytest-django
pytest-factoryboy
pytest-asyncio